backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.4.2
websockets==11.0.3
python-multipart==0.0.6
pillow==10.1.0
aiofiles==23.2.1 